requirements.txt
aiohttp==0.22.5
beautifulsoup4==4.4.1
chardet==2.3.0
colorama==0.3.7
discord.py==0.10.0
dnspython3==1.12.0
docutils==0.12
IPy==0.83
Jinja2==2.8
MarkupSafe==0.23
multidict==1.2.2
mysqlclient==1.3.7
Pillow==3.1.1
py==1.4.31
pyfiglet==0.7.4
pytest==2.9.1
python-valve==0.1.1
qrcode==5.2.2
requests==2.9.1
six==1.10.0
SQLAlchemy==1.0.12
websockets==3.1
ws4py==0.3.4
